Fix many warnings when building the program on Windows with Visual Studio.
git-svn-id: http://svn.powertabs.org/svn/ptedev/trunk@1185 5b1460fe-822e-4aa1-ac1b-8a6412c92c8c

+++ source/app/caret.cpp
@@ -97,8 +97,8 @@ void Caret::moveToLastSystem()
 void Caret::moveStaff(int offset)
 {
     myLocation.setStaffIndex(boost::algorithm::clamp(
-            myLocation.getStaffIndex() + offset, 0,
-            myLocation.getSystem().getStaves().size() - 1));
+        myLocation.getStaffIndex() + offset, 0,
+        static_cast<int>(myLocation.getSystem().getStaves().size() - 1)));
 
     onLocationChanged();
 }
@@ -164,7 +164,7 @@ int Caret::getLastPosition() const
 
 int Caret::getLastSystemIndex() const
 {
-    return myLocation.getScore().getSystems().size() - 1;
+    return static_cast<int>(myLocation.getScore().getSystems().size() - 1);
 }
 
 void Caret::moveToPosition(int position)
@@ -189,8 +189,8 @@ void Caret::moveToSystem(int system, bool keepStaff)
         else
         {
             myLocation.setStaffIndex(boost::algorithm::clamp(
-                    myLocation.getStaffIndex(), 0,
-                    myLocation.getSystem().getStaves().size()));
+                myLocation.getStaffIndex(), 0,
+                static_cast<int>(myLocation.getSystem().getStaves().size())));
         }
 
         myLocation.setPositionIndex(0);
